+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 ### Bug Fixes
 
 * ch.pipline 顺序问题会导致加解码类失效 ([30e1a55](https://github.com/ctywzy/MyDubbo/commit/30e1a559a2728ebd373ee9381aa41b87754d7830))
+* 对象没有实现序列化接口导致转成字节数组失败 ([060b744](https://github.com/ctywzy/MyDubbo/commit/060b744addf431db90402d64dbddb8cfbe9d4e31))
 
 
 ### Features
 
 * changeLog 功能添加 ([a8e3cec](https://github.com/ctywzy/MyDubbo/commit/a8e3cec532f00a8ec5908ed74b82ed0af9906c08))
+* 利用netty的Bufbyte进行数据传输 ([4772ff1](https://github.com/ctywzy/MyDubbo/commit/4772ff1993bc9e35a3aa4e26cd5efdac8bde4f32))
 * 客户端调服务端 ([ef4e48e](https://github.com/ctywzy/MyDubbo/commit/ef4e48e2bf1e1772d47bbad38e550ca974779246))
 * 由自动连接调用改为主动调用 ([ae096e9](https://github.com/ctywzy/MyDubbo/commit/ae096e9732b475329bbf4f6ed803ff97e89f3d68))
 * 给客户端和服务器增加 加解码操作使之可以通信 ([531d85b](https://github.com/ctywzy/MyDubbo/commit/531d85b300b512cb0b31a8885fb949c88a08cbd4))